EUI-2454 refactored navigation
