+++ src/noc/containers/noc-home/noc-home.component.html
@@ -2,11 +2,12 @@
 <div class="govuk-width-container">
   <exui-page-wrapper>
     <exui-noc-navigation
-      (back)="back($event)"
-      (continue)="continue($event)"
-      (submit)="submit($event)"
+      (onEvent)="onNavEvent($event)"
     >
-      <exui-noc-case-ref *ngIf="isComponentVisible((nocNavigationCurrentState$ | async), nocState.START)"></exui-noc-case-ref>
+      <exui-noc-case-ref
+        *ngIf="isComponentVisible((nocNavigationCurrentState$ | async), nocState.START)"
+        [navEvent]="navEvent"
+      ></exui-noc-case-ref>
     </exui-noc-navigation>
   </exui-page-wrapper>
 </div>
