+++ src/noc/containers/noc-navigation/noc-navigation.component.html
@@ -1,6 +1,6 @@
-<a class="govuk-back-link" href="javascript:void(0)" *ngIf="isVisible((nocNavigationCurrentState$ | async), backVisibilityStates)" (click)="onBack($event)">Back</a>
+<a class="govuk-back-link" href="javascript:void(0)" *ngIf="isVisible((nocNavigationCurrentState$ | async), backVisibilityStates)" (click)="onEventTrigger(nocNavigationEvent.BACK)">Back</a>
 <div>
     <ng-content></ng-content>
 </div>
-<button class="button" type="button" *ngIf="isVisible((nocNavigationCurrentState$ | async), continueVisibilityStates)" (click)="onContinue($event)">Continue</button>
-<button class="button" type="button" *ngIf="isVisible((nocNavigationCurrentState$ | async), submitVisibilityStates)" (click)="onSubmit($event)">Submit</button>
+<button class="button" type="button" *ngIf="isVisible((nocNavigationCurrentState$ | async), continueVisibilityStates)" (click)="onEventTrigger(nocNavigationEvent.CONTINUE)">Continue</button>
+<button class="button" type="button" *ngIf="isVisible((nocNavigationCurrentState$ | async), submitVisibilityStates)" (click)="onEventTrigger(nocNavigationEvent.SUBMIT)">Submit</button>
