+++ src/noc/models/noc-answer.interface.ts
@@ -1,0 +1,4 @@
+export interface NocAnswer {
+    displayOrder: number;
+    answer: string;
+}
