+++ src/noc/models/noc-error.interface.ts
@@ -1,0 +1,4 @@
+export interface NocError {
+    responseCode: number;
+    message: string;
+}
