+++ src/noc/models/noc-event.interface.ts
@@ -1,0 +1,6 @@
+import { NocAnswer } from './noc-answer.interface';
+
+export interface NocEvent {
+    caseReference: string;
+    nocAnswers: NocAnswer[];
+}
