+++ src/noc/models/noc-navigation-event.enum.ts
@@ -1,0 +1,5 @@
+export enum NocNavigationEvent {
+    BACK,
+    CONTINUE,
+    SUBMIT
+}
