+++ src/noc/models/noc-options.interface.ts
@@ -1,0 +1,4 @@
+export interface NocOptions {
+    askForReason: boolean;
+    showRequestType: boolean;
+}
