+++ src/noc/models/noc-question.interface.ts
@@ -1,0 +1,6 @@
+export interface NocQuestion {
+    displayOrder: number;
+    answerType: string;
+    displayContext: string;
+    questionLabel: string;
+}
