+++ src/noc/models/noc-state.enum.ts
@@ -1,0 +1,13 @@
+export enum NocState {
+    START,
+    CASE_REF_VALIDATION_FAILURE,
+    CASE_REF_SUBMISSION_FAILURE,
+    QUESTION,
+    ANSWER_INCOMPLETE,
+    ANSWER_SUBMISSION_FAILURE,
+    CHECK_ANSWERS,
+    AFFIRMATION_NOT_AGREED,
+    SUBMISSION_SUCCESS_APPROVED,
+    SUBMISSION_SUCCESS_PENDING,
+    SUBMISSION_FAILURE
+}
