+++ src/noc/store/reducers/index.ts
@@ -1,5 +1,5 @@
 import {ActionReducerMap, createFeatureSelector} from '@ngrx/store';
-import { NocStateData } from '../../models/noc.state';
+import { NocStateData } from '../../models/noc-state-data.interface';
 
 import * as fromNoc from './noc.reducer';
