+++ src/noc/store/reducers/noc.reducer.spec.ts
@@ -1,5 +1,5 @@
+import { NocState } from 'src/noc/models/noc-state.enum';
 import * as fromActions from '../actions/noc.action';
-import { NocState } from '../../models/noc.state';
 import * as fromReducer from './noc.reducer';
 
 describe('Noc Navigation Reducer', () => {
