+++ src/noc/store/reducers/noc.reducer.ts
@@ -1,5 +1,6 @@
 import * as fromActions from '../actions';
-import { NocState, NocStateData } from '../../models/noc.state';
+import { NocStateData } from '../../models/noc-state-data.interface';
+import { NocState } from 'src/noc/models/noc-state.enum';
 
 export const initialState: NocStateData = {
     state: NocState.START,
